[BUGFIX] The activity could be recreated during background mode.
When camera preview is on TV in background mode the activity of the application
could be destroyed by the system (mostly because of the lack or resources
required by other apps). In this situation activity membors have to be
properly reinitialized. Camera preview will be restarted to pass a new instance
of SurfaceHolder/Surface to mediaserver.